Moving freeriding attacks to their own module.
